Integrated Global Network Graph into Index.py
