+++ horus/apps/local_network_graph.py
@@ -20,7 +20,7 @@ from flask import (
     jsonify,
 )
 
-from app import app
+from index import app
 from data_analysis_module.network_graph import diogenetGraph
 
 dict_of_datasets = {'Diogenes Laertius': 'diogenes', 'Life of Pythagoras Iamblichus': 'iamblichus'}
